Break up everyting into separate files, reorganize file structure
