+++ server/src/resources/budget/budget.model.js
@@ -1,27 +1,29 @@
-const { DataTypes } = require('sequelize');
-
-const { sequelize } = require('../../utils/db');
-const UserModel = require('../user/user.model');
-
-const Budget = sequelize.define('budget', {
-  userId: {
-    type: DataTypes.INTEGER,
-    references: {
-      model: UserModel,
-      key: 'id'
+module.exports = (sequelize, DataTypes) => {
+  const budgetModel = sequelize.define('budget', {
+    userId: {
+      type: DataTypes.INTEGER,
+      references: {
+        model: 'user',
+        key: 'id'
+      },
+      field: 'user_id'
+    },
+    agency: {
+      type: DataTypes.STRING
+    },
+    monthlyPassCost: {
+      type: DataTypes.STRING,
+      field: 'monthly_pass_cost'
+    },
+    fareCost: {
+      type: DataTypes.STRING,
+      field: 'fare_cost'
     }
-  },
-  agency: {
-    type: DataTypes.STRING
-  },
-  monthlyPassCost: {
-    type: DataTypes.STRING
-  },
-  fareCost: {
-    type: DataTypes.STRING
-  }
-});
+  });
 
-Budget.belongsTo(UserModel, {});
+  budgetModel.associate = models => {
+    budgetModel.belongsTo(models.user, { foreignKey: 'id', as: 'user' });
+  };
 
-module.exports = Budget;
+  return budgetModel;
+};
